Add new macro to print tokens resistanced
The macro the condition immunities, damage immunities,
damage resistances and damage vulnerabilities of the
currently selected token(s). Damage types that appeared
in the previous chat message (e.g. due to a roll) are
highlighted in red. A DM can call this macro after a
player rolled damage to quickly see if they need to
apply full, half or double damage.