Create create_merchant_stock
This macro will generate stock on an actor great if you use Loot NPC where you can create a merchant inventory. 
This will take items from an item compendium and fill random items in the inventory. The quantity is random based on the roll formula you use for how many items should be picked. It does not check for duplicates. 
If the price is 0 It will use the merchant price calculator defined in XGtE