Create initiativeWithDisadvantage.js
Initiative with Disadvantage by Nulmas#9462
Thanks to Freeze#2689 and vance#1935 for the help.

This macro allows GMs and players to roll for Initiative with disadvantage when playing D&D 5e. Hopefully it won't be needed for long and the option for it will be added to the system in a future release.

The macro will roll for all the selected tokens and add them to the combat if they aren't in it already. It will also check if you are using Dex as a tiebreaker and roll accordingly.

BEWARE: If a token has already rolled for initiative and you use this macro with it selected, the new initiative will replace the old one. I considered changing this, but decided it's worth keeping it this way in case a player or GM rolls for initiative without disadvantage by mistake.