Create auto_sort_creatures_by_type.js
Created a macro to automatically sort all created creatures by their creature type. This macro creates folders in the root of the actors directory and the adds the actor fitting to this creature type.
Very helpful when there are a lot of actors and you need to get a better overview as DM.